% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus_index.R
\name{amtl_ratio}
\alias{amtl_ratio}
\title{Calculate antemortem tooth loss}
\usage{
amtl_ratio(.data, .amtl, ...)
}
\arguments{
\item{.data}{Data frame in long format, with one tooth surface per row, and one calculus score per surface.}

\item{.amtl}{AMTL count variable.}

\item{...}{Can be used to add grouping variable(s). Group by tooth, tooth type, individual, etc.}
}
\value{
Returns a data frame with amtl ratio per group.
}
\description{
Calculate antemortem tooth loss
}
\details{
Calculate the caries ratio for each tooth type. Calculated as the number
of caries divided by the number of visible surfaces.
}
