% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calculus_full}
\alias{calculus_full}
\title{Middenbeemster dental calculus data}
\format{
\subsection{\code{calculus_full}}{

A data frame with 41 rows and 97 columns:
\describe{
\item{id}{Skeletal ID}
\item{t11:t48_buc/lin/ipx}{An integer denoting the calculus score (0-3) on
a given tooth surface: tooth number (FDI notation)_surface.
buc = buccal; lin = lingual, ipx = inter-proximal.}
...
}
}
}
\source{
https://doi.org/10.5281/zenodo.8010630
}
\usage{
calculus_full
}
\description{
Dental calculus scores collected from Middenbeemster (MB11), a rural 19th century
archaeological site in the Netherlands.
}
\keyword{datasets}
