% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus_index.R
\name{calculus_index}
\alias{calculus_index}
\title{Calculate calculus index}
\usage{
calculus_index(.data, simple = FALSE)
}
\arguments{
\item{.data}{Data frame in long format, with one tooth surface per row, and one calculus score per surface.}
}
\value{
Returns a data frame with four rows per individual containing the calculated calculus
index for each quadrant (CI-UP, CI-LP, CI-UA, CI-LA).
}
\description{
Calculate calculus index
}
\details{
Calculate the calculus index for each quadrant of an individual.
Based on Greene et al. 2005.
}
