% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caries}
\alias{caries}
\title{Middenbeemster dental caries data}
\format{
\subsection{\code{caries}}{

A data frame with 41 rows and 33 columns:
\describe{
\item{id}{Skeletal ID}
\item{t11:t48}{a character string denoting the location of a caries lesion
on a given tooth (FDI notation), with multiple lesions separated by \verb{;}}
}
}
}
\source{
https://doi.org/10.5281/zenodo.8010630
}
\usage{
caries
}
\description{
Dental caries scores collected from Middenbeemster (MB11), a rural 19th century
archaeological site in the Netherlands.
}
\keyword{datasets}
