% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus_index.R
\name{caries_ratio}
\alias{caries_ratio}
\title{Calculate caries ratio}
\usage{
caries_ratio(.data, .caries, ...)
}
\arguments{
\item{.data}{Data frame in long format, with one tooth surface per row, and one calculus score per surface.}

\item{.caries}{Caries count variable.}

\item{...}{Can be used to add grouping variable(s). Group by tooth, tooth type, individual, etc.}
}
\value{
Returns a data frame with caries ratio per group.
}
\description{
Calculate caries ratio
}
\details{
Calculate the caries ratio for each tooth type. Calculated as the number
of caries divided by the number of visible surfaces.
}
