% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor_statement}
\alias{cor_statement}
\title{Function to convert compound abbreviations to full name
correlation statements}
\usage{
cor_statement(.data, strength)
}
\arguments{
\item{.data}{correlation tibble. Long format.}

\item{strength}{character. Strength of correlation ("weak", "moderate", or "strong").}
}
\description{
Function to convert compound abbreviations to full name
correlation statements
}
