% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dental_inv}
\alias{dental_inv}
\title{Middenbeemster dental inventory}
\format{
\subsection{\code{dental_inv}}{

A data frame with 41 rows and 33 columns:
\describe{
\item{id}{Skeletal ID}
\item{t11:t48}{a character string denoting the status of a tooth.
p = present, m = missing (for unknown reason), aml = antemortem loss, dna = tooth missing because it is being sampled for DNA.}
}
}
}
\source{
https://doi.org/10.5281/zenodo.7649151
}
\usage{
dental_inv
}
\description{
Data on dental status.
}
\keyword{datasets}
