% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dental_longer}
\alias{dental_longer}
\title{Convert wide data frame containing dental info to long format and add tooth
position variables.}
\usage{
dental_longer(.data, cols)
}
\arguments{
\item{.data}{A data frame containing one column with a unique identifier and
multiple columns with calculus scores from each surface of a tooth.}

\item{cols}{Which columns to pivot into longer format. These should be the columns
containing the scores for each tooth.}
}
\value{
Returns a data frame with tooth number, region (maxilla, mandible),
position (anterior, posterior), side, tooth_type (incisor, canine, etc.)
}
\description{
Convert wide data frame containing dental info to long format and add tooth
position variables.
}
