% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dental-plot.R
\name{dental_plot}
\alias{dental_plot}
\title{Plotting dental features}
\usage{
dental_plot(.data, fill, notation)
}
\arguments{
\item{.data}{Data frame containing...}

\item{fill}{What variable is used for the 'fill' argument. Passed to ggplot().}

\item{notation}{Currently not in use. May be used to support different notation in the future.}
}
\description{
Plotting dental features
}
