% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dentition_checker}
\alias{dentition_checker}
\title{Dentition checker}
\usage{
dentition_checker(.data)
}
\arguments{
\item{.data}{A data frame containing one column with a unique identifier and
multiple columns with calculus scores from each surface of a tooth.}
}
\value{
Returns a warning if data frame does not fulfill criterion.
}
\description{
Checks if the data frame satisfies the criterion that all three surfaces must
be present between both sides of the dentition.
}
