% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Sample metadata}
\format{
\subsection{\code{metadata}}{

A data frame with 44 rows and 8 columns:
\describe{
\item{id}{Skeletal ID}
\item{sample}{Dental calculus sample number}
\item{element}{FDI notation for tooth/teeth sampled for each individual.}
\item{KZ_sample}{Dental calculus sample number on tubes from previous study.}
\item{KZ_element}{FDI notation for tooth/teeth sampled from the previous study for each individual.}
\item{replicated}{Whether or not the sample was replicated in a separate UHPLC-MS/MS analysis.}
\item{batch1_weight}{weight (mg) of dental calculus sample analysed in the first batch.}
\item{batch2_weight}{weight (mg) of dental calculus sample analysed in the replication batch.}
...
}
}
}
\source{
https://doi.org/10.5281/zenodo.8010630
}
\usage{
metadata
}
\description{
Metadata for the samples analysed by UHPLC-MS/MS.
}
\keyword{datasets}
