% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{path_cond_clean}
\alias{path_cond_clean}
\title{Middenbeemster pathological conditions}
\format{
\subsection{\code{path_cond_clean}}{

A data frame with 41 rows and 13 columns:
\describe{
\item{id}{Skeletal ID}
\item{OA}{Presence (TRUE) or absence (FALSE) of lesions associated with osteoarthritis.}
\item{IVDD}{Presence (TRUE) or absence (FALSE) of lesions associated with intervertebral disc disease.}
\item{TB}{Presence (TRUE) or absence (FALSE) of lesions associated with tuberculosis.}
\item{Mastoiditis}{Presence (TRUE) or absence (FALSE) of lesions associated with mastoiditis.}
\item{DISH}{Presence (TRUE) or absence (FALSE) of lesions associated with DISH.}
\item{VOP}{Presence (TRUE) or absence (FALSE) of lesions associated with vertebral osteophytosis.}
\item{SN}{Presence (TRUE) or absence (FALSE) of lesions associated with Schmorl's nodes.}
\item{DDD}{Presence (TRUE) or absence (FALSE) of lesions associated with degenerative disc disease.}
\item{PNBF}{Presence (TRUE) or absence (FALSE) of periosteal new bone formation.}
\item{OD}{Presence (TRUE) or absence (FALSE) of lesions associated with osteochondritis dissecans.}
\item{CF}{Presence (TRUE) or absence (FALSE) of cribra femora.}
\item{CO}{Presence (TRUE) or absence (FALSE) of cribra orbitalia.}
}
}
}
\source{
https://doi.org/10.5281/zenodo.7649151
}
\usage{
path_cond_clean
}
\description{
Data on various disease diagnoses of individuals from Middenbeemster based on skeletal lesions.
}
\keyword{datasets}
