% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{periap}
\alias{periap}
\title{Middenbeemster periapical lesion data}
\format{
\subsection{\code{periap}}{

A data frame with 41 rows and 33 columns:
\describe{
\item{id}{Skeletal ID}
\item{t11:t48}{a character string denoting the location of a periapical lesion
on a given tooth (FDI notation), with multiple lesions separated by \verb{;}}
}
}
}
\source{
https://doi.org/10.5281/zenodo.8010630
}
\usage{
periap
}
\description{
Data on periapical lesions collected from Middenbeemster (MB11), a rural 19th century
archaeological site in the Netherlands.
}
\keyword{datasets}
