% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{periodont}
\alias{periodont}
\title{Middenbeemster periodontitis data}
\format{
\subsection{\code{periodont}}{

A data frame with 41 rows and 33 columns:
\describe{
\item{id}{Skeletal ID}
\item{t11:t48}{an integer denoting periodontitis score (0-3) for a given tooth (FDI notation)}
}
}
}
\source{
https://doi.org/10.5281/zenodo.8010630
}
\usage{
periodont
}
\description{
Periodontitis scores collected from Middenbeemster (MB11), a rural 19th century
archaeological site in the Netherlands. Scores were recorded on a scale of
0-3 based on amount of alveolar bone loss.
}
\keyword{datasets}
