% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sinusitis_clean}
\alias{sinusitis_clean}
\title{Middenbeemster sinusitis}
\format{
\subsection{\code{sinusitis_clean}}{

A data frame with 42 rows and 3 columns:
\describe{
\item{id}{Skeletal ID}
\item{CMS}{Presence (TRUE) or absence (FALSE) of lesions associated with chronic maxillary sinusitis.}
\item{IPR}{Presence (TRUE) or absence (FALSE) of periosteal reaction on visceral surface of ribs.}
}
}
}
\source{
https://doi.org/10.5281/zenodo.7649151
}
\usage{
sinusitis_clean
}
\description{
Data on sinusitis diagnosis of individuals from Middenbeemster based on skeletal lesions.
}
\keyword{datasets}
