% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tooth_position}
\alias{tooth_position}
\title{Produce detailed tooth region information from tooth number (FDI notation)}
\usage{
tooth_position(.data)
}
\arguments{
\item{.data}{data frame containing tooth number (t + FDI notation)}
}
\value{
Returns a data frame with tooth number, region (maxilla, mandible),
position (anterior, posterior), side, tooth_type (incisor, canine, etc.)
}
\description{
Produce detailed tooth region information from tooth number (FDI notation)
}
