% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uhplc_data_comb}
\alias{uhplc_data_comb}
\title{UHPLC analysis data from both batches}
\format{
\subsection{\code{uhplc_data_comb}}{

A data frame with 42 rows and 91 columns:
\describe{
\item{sample}{number on the sample tube.}
\item{\if{html}{\out{<compound>}}_wash{1..3}_batch{1..2}}{Absolute quantity (ng) of compound detected in sample washes.}
\item{\if{html}{\out{<compound>}}_calc_batch{1..2}}{Absolute quantity (ng) of compound detected after sample dissolution.}
\item{batch{1..2}_weight}{Weight of the calculus sample in mg.}
...
}
}
}
\source{
https://doi.org/10.5281/zenodo.8010630
}
\usage{
uhplc_data_comb
}
\description{
Contains the results for each sample from each extraction step (washes, and dissolution)
analysed by UHPLC in the first and second (replication) batch.
}
\keyword{datasets}
