% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor_statement}
\alias{cor_statement}
\title{Function to convert compound abbreviations to full name
correlation statements}
\usage{
cor_statement(.data, cor_coef, strength = c("weak", "moderate", "strong"))
}
\description{
Function to convert compound abbreviations to full name
correlation statements
}
