% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lloq}
\alias{lloq}
\title{Lower limits of quantitation (LLOQ)}
\format{
\subsection{\code{lloq}}{

A data frame with 76 rows and 2 columns:
\describe{
\item{compound}{Name of the compound.}
\item{lloq}{Lower limit of quantitation (ng)}
...
}
}
}
\source{
https://doi.org/10.5281/zenodo.8010630
}
\usage{
lloq
}
\description{
LLOQ values for the targeted compounds.
}
\keyword{datasets}
